audio_detector/requirements_all.txt
antlr4-python3-runtime==4.8
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
audioread==3.0.1
bitarray==2.9.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
cython==3.0.9
decorator==5.1.1
exceptiongroup==1.2.0
# fairseq==0.12.2
filelock==3.13.3
fonttools==4.51.0
fsspec==2024.3.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
hydra-core==1.0.7
idna==3.6
importlib-metadata==7.0.2
importlib-resources==6.3.0
inflection==0.5.1
itsdangerous==2.1.2
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
librosa==0.9.2
llvmlite==0.41.1
lxml==5.1.0
markupsafe==2.1.5
matplotlib==3.7.5
numba==0.58.1
numpy==1.24.4
omegaconf==2.0.6
packaging==24.0
pandas==2.0.3
pillow==10.2.0
pkgutil-resolve-name==1.3.10
platformdirs==4.2.0
pooch==1.8.1
portalocker==2.8.2
protobuf==5.26.0
pycparser==2.21
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
pyyaml==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
resampy==0.4.3
rpds-py==0.18.0
sacrebleu==2.4.1
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
starlette==0.37.2
tabulate==0.9.0
tensorboardx==2.6.2.2
threadpoolctl==3.3.0
torchaudio==0.13.1
tokenizers==0.15.2
tqdm==4.66.2
transformers==4.39.3
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.18.1
connexion==3.0.0
connexion[uvicorn]
Flask==2.3.2
a2wsgi==1.10.6
fastapi==0.111.0
pyjwt==2.8.0
jinja2==3.1.4

